Fix a lifecycle-related bug on ExecutionActivity
The bug was that if a user makes the activity back-ground when an execution is running,
then the app crushes because of an IllegalStateException which comes from
the AlertDialogFragment in onComplete method.
In that case Activity#onStop is already called when onComplete is
processing.

To resolve this, added a setter method for AlertDialogFragment#Builder
to specify which commit method (commit/commitAllowingStateLoss) is used
for showing the fragment.