Standardize controllers (#59)
This commit standardizes the MachineController to make it has the union of its children's methods.
We keep MachineController as it were and add 'CarControllerBase'
for car-formed machines such as the car form of Nxt / Ev3 and Pile robots,
and rename NxtController after it (NxtCarController).
By this change, the previous block classes and resources related to the car-formed machines
can be moved in 'main' variant and we decide to keep them in 'car' package for readability
(StopForSecBlock is not moved because it doesn't depend on the car-formed machines' method).

This commit also cleans string resources.

* TL;DR
** Background
We talked over the design of controllers again and again offline, and we had two options;

- Make controllers for each machine (Nxt, Ev3, Pile, etc.) so that they have their specific methods
(e.g., Pile Robot doesn't have ColorSensor, thus it doesn't have getColorSensorValue or something like that)
and keep the MachineController as it is.
- Make the common MachineController that has the union of methods of these controllers.

They actually have pros & cons and our opinions conflicted. But finally, at present, we decided to choose the latter.

** Pros & Cons of the former
*** Pros
- Straightforward way and thus readable. This is very important for developers.
- They can be easily tested.

*** Cons
- There will be many duplications of codes because these controllers use [our library](https://github.com/PileProject/drivecommand)'s
MachineBase that provides the same interface to control machines (so, we will have very similar codes in blocks like
[ForwardSecBlock](https://github.com/PileProject/drive/blob/develop/app/src/nxt/java/com/pileproject/drive/programming/visual/block/sequence/ForwardSecBlock.java)
in each build variant) but we currently can't make the most of it because of the design.
- It doesn't follow the design of [our library](https://github.com/PileProject/drivecommand). But we can fix that of our library easily.

** Pros & Cons of the latter
*** Pros
- It follows the design of [our library](https://github.com/PileProject/drivecommand).
- It commonizes the blocks through MachineController interface. Thus, it reduces the duplicated codes not only *.java but also *.xml.
- Easy to maintain because the codes are unified.

*** Cons
- It is difficult to test each variant.
- Not straightforward way.
- Users can't know what kinds of methods each controller has, and make it worse,
calling wrong methods lead to RuntimeError not CompileError.
- Poor extensibility. We should add/remove a method to/from MachineController itself if new machines with new methods.
So, maybe we should make an another layer for other machines something
like MachineController -> CarController, DroneController (i.e., not use the MachineController as the current NxtController's parent).

** Why we chose the latter
This was a very difficult problem for us but we took the pros of the latter.
Maintaining duplicated codes *.java and *.xml is very annoying
and we can easily change our mind when a new type of machine comes
if we keep the unified codes clean.