Clean up comments (#62)
Solves #24. This is a PR to clean up comments/javadocs. This may be related to #56.

# Background
Before moving on to adding other variants (ev3/pile), I think we should fix some comments/javadocs because we should do the same changes to the other variants if we implement them. At present, some javadocs have errors with their grammar.


# Goal
The goal is not enriching but cleaning comments up. Of course, I add some information but they may be not enough. 


# Note
I add a plugin to generate javadocs automatically. It introduces some tasks like `generateNxtDebugJavadoc`. Running these tasks causes some warnings because of the reference errors of auto-generated classes but we can see the complete Javadocs in `app/javaDoc`. I believe that the document is really helpful not only for others but also for us.

However, I'm sorry that we should care about the tasks so that they shouldn't have errors related to auto-generated classes. But please don't worry about that. I will check them when you submit a new PR.